+++ src/neural_network/layer/neuron/RecurrentNeuron.cpp
@@ -22,7 +22,7 @@ float RecurrentNeuron::output(const vector<float>& inputs, bool temporalReset)
     this->previousOutput = this->lastOutput;
     this->sum = 0;
     int w;
-    for (w = 0; w < inputs.size(); ++w)
+    for (w = 0; w < (int)inputs.size(); ++w)
     {
         this->sum += inputs[w] * this->weights[w];
     }
@@ -55,7 +55,7 @@ inline
 void RecurrentNeuron::updateWeights(const float error)
 {
     int w;
-    for (w = 0; w < this->lastInputs.size(); ++w)
+    for (w = 0; w < (int)this->lastInputs.size(); ++w)
     {
         auto deltaWeights = this->optimizer->learningRate * error * this->lastInputs[w];
         deltaWeights += this->optimizer->momentum * this->previousDeltaWeights[w];
