+++ tests/dataset_tests/Iris/IrisTest.cpp
@@ -29,10 +29,10 @@ TEST_F(IrisTest, loadData)
 {
     ASSERT_EQ(data->sizeOfData, 4);
     ASSERT_EQ(data->numberOfLabel, 3);
-    ASSERT_EQ(data->sets[training].inputs.size(), 150);
-    ASSERT_EQ(data->sets[training].labels.size(), 150);
-    ASSERT_EQ(data->sets[snn::testing].inputs.size(), 150);
-    ASSERT_EQ(data->sets[snn::testing].labels.size(), 150);
+    ASSERT_EQ((int)data->sets[training].inputs.size(), 150);
+    ASSERT_EQ((int)data->sets[training].labels.size(), 150);
+    ASSERT_EQ((int)data->sets[snn::testing].inputs.size(), 150);
+    ASSERT_EQ((int)data->sets[snn::testing].labels.size(), 150);
     ASSERT_EQ(data->sets[snn::testing].numberOfTemporalSequence, 0);
     ASSERT_EQ(data->sets[snn::testing].numberOfTemporalSequence, 0);
     ASSERT_EQ(data->isValid(), 0);
@@ -50,5 +50,5 @@ TEST_F(IrisTest, trainNeuralNetwork)
     neuralNetwork.waitFor(0.98_acc || 2_s);
     neuralNetwork.stopTraining();
     auto accuracy = neuralNetwork.getGlobalClusteringRate();
-    ASSERT_ACCURACY(accuracy, 0.98);
+    ASSERT_ACCURACY(accuracy, 0.98f);
 }
