+++ tests/dataset_tests/daily-min-temperatures/DailyMinTemperaturesTest.cpp
@@ -29,10 +29,10 @@ TEST_F(DailyMinTemperaturesTest, loadData)
 {
     ASSERT_EQ(data->sizeOfData, 1);
     ASSERT_EQ(data->numberOfLabel, 1);
-    ASSERT_EQ(data->sets[training].inputs.size(), 3649);
-    ASSERT_EQ(data->sets[training].labels.size(), 3649);
-    ASSERT_EQ(data->sets[snn::testing].inputs.size(), 3649);
-    ASSERT_EQ(data->sets[snn::testing].labels.size(), 3649);
+    ASSERT_EQ((int)data->sets[training].inputs.size(), 3649);
+    ASSERT_EQ((int)data->sets[training].labels.size(), 3649);
+    ASSERT_EQ((int)data->sets[snn::testing].inputs.size(), 3649);
+    ASSERT_EQ((int)data->sets[snn::testing].labels.size(), 3649);
     ASSERT_EQ(data->sets[snn::testing].numberOfTemporalSequence, 1);
     ASSERT_EQ(data->sets[snn::testing].numberOfTemporalSequence, 1);
     ASSERT_EQ(data->numberOfLabel, 1);
